Movement update
In this update the map function was updated a bit but more importantly movement and an the beginnings of an interface was added to the programing allowing the movement of the player character (which is planned to become an object in future updates). The movement function allows one to go to an adjacent room if a tunnel is connecting the rooms together and gives the description of the room upon entering it.