Obstacles update
Obstacles/barriers have been added to the code. The obstacles can be divided into three categories, keypad obstacles(input correct number to progress), key item obstacle (certain item required in inventory to progress) and movable/breakable obstacle (the name is self-explanatory). The code to interact with and solve obstacles is only done for "keypad obstacles" and the rest of the code will be finished in the next update.